Move test_for_any_all to functional
