+++ tests/functional/ext/for_any_all/for_any_all.py
@@ -59,3 +59,55 @@ def words_contains_word(words):
         if word == "word":
             return True
     return False
+
+
+def no_suggestion_if_not_if():
+    """Do not emit if the for loop does not have the pattern we are looking for"""
+    for val in range(1):
+        var = val
+        return var
+
+def no_suggestion_if_not_bool(item):
+    """Do not emit if the if-statement does not return a bool"""
+    for parent in item.parents():
+        if isinstance(parent, str):
+            return "True"
+    return "False"
+
+def print_items(items):
+    """Do not emit if there is no If condition in the for loop."""
+    for item in items:
+        print(item)
+    return True
+
+def print_items2(items):
+    """Do not emit if anything besides a boolean is returned."""
+    for item in items:
+        return item
+    return True
+
+def print_items3(items):
+    """Do not emit if anything besides a boolean is returned."""
+    for _ in items:
+        return False
+    return items
+
+def print_items4(items):
+    """Do not emit if there is more logic which can cause side effects
+    or become less readable in a list comprehension.
+    """
+    for item in items:
+        if isinstance(item, str):
+            print(item)
+            return False
+    return True
+
+def is_from_decorator(node):
+    """Do not emit if the if has an else condition. Generally implies more complicated logic."""
+    for parent in node.node_ancestors():
+        if isinstance(parent, str): # pylint: disable=no-else-return
+            return True
+        else:
+            if parent in parent.selected_annotations:
+                return False
+            return False
